general utils moved to helpers, corresponding refactorings, function def formatting
