+++ src/helpers.js
@@ -11,6 +11,9 @@ governing permissions and limitations under the License.
 const querystring = require('querystring')
 const httplinkheader = require('http-link-header')
 const url = require('url')
+const loggerNamespace = '@adobe/aio-lib-events'
+const logger = require('@adobe/aio-lib-core-logging')(loggerNamespace,
+    { level: process.env.LOG_LEVEL })
 
 /**
  * Reduce an Error to a string
@@ -81,9 +84,63 @@ function parseRetryAfterHeader (header) {
   }
 }
 
+/**
+ * Wrapper to check the event received by webhook
+ * and return decoded (if encoded) and properly parsed payload
+ *
+ * @param {*} event event payload received by webhook
+ * @returns decoded and properly parsed payload
+ */
+function getProperPayload (event) {
+  let decodedJsonPayload
+  try {
+      if (isBase64Encoded(event)) {
+          decodedJsonPayload = JSON.parse(Buffer.from(event, 'base64').toString('utf-8'))
+      } else {
+          // parsing for non-encoded json payloads (e.g. custom events)
+          decodedJsonPayload = JSON.parse(event)
+      }
+  } catch (error) {
+      logger.error('error occured while checking payload' + error.message)
+      return genErrorResponse(400, 'Failed to understand the payload')
+  }
+  return decodedJsonPayload
+}
+
+/**
+ * Checks for the payload type if it is base64 encode or not
+ * 
+ * @param {*} eventPayload - the event payload received by the consumer webhook
+ * @returns {boolean} true if encoded or false
+ */
+function isBase64Encoded (eventPayload) {
+  return Buffer.from(eventPayload, 'base64').toString('base64') === eventPayload
+}
+
+/**
+ * Generates generic error json response object based on HTTP error codes and message
+ * 
+ * @param {*} statusCode HTTP error codes
+ * @param {*} message custom error message
+ */
+function genErrorResponse (statusCode, message) {
+  const response = {
+      statusCode: statusCode,
+      body: message,
+      headers: {
+          'Content-Type': 'application/json'
+      }
+  }
+  return {
+      error: response
+  }
+}
+
 module.exports = {
   reduceError,
   appendQueryParams,
   parseLinkHeader,
-  parseRetryAfterHeader
+  parseRetryAfterHeader,
+  getProperPayload,
+  genErrorResponse
 }
